% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloha.R
\name{say_aloha}
\alias{say_aloha}
\title{Title}
\usage{
say_aloha(name, print = TRUE)
}
\arguments{
\item{name}{a character}

\item{print}{a logical}
}
\value{
a character
}
\description{
Title
}
\examples{
say_aloha("Allison")
}
